✨(project) import the library from ashley
The library has been extracted from the `openfun/ashley` repository in
order to use it in multiple projects. It has been renamed (the
original package name was `lti_provider`) in order to avoid confusion
with other LTI related projects.

This is basically a copy of the work done in Ashley, with our common
build process and quality check tools. The views have been simplified
and a sandbox has been added to try out the project easily.