♻️(lti_toolbox) route captured URL value to `_do_on_success`
While working with the URL dispatcher, it was necessary to capture a value
from the URL using angle brackets. This  captured value was channeled
as an argument into the `post` method.

This method serves as a bridge to the corresponding abstract method,
and I required a seamless way to transmit this captured value from the
`post method` to the final step, denoted as `_do_on_success`.